backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
firebase-admin==6.3.0
google-cloud-aiplatform==1.71.0
google-cloud-storage==2.14.0
PyPDF2==3.0.1
python-docx==1.1.0
gunicorn==21.2.0
